// src/types.ts
export interface EventData {
  id: string;
  title: string;
  startTime: string;
  endTime: string;
  description?: string;
  color?: string;
  date?: string;
}

export interface Event {
  id: string;
  title: string;
  startTime: string;
  endTime: string;
  description?: string;
  color?: string;
  date?: string;
}